Move custom_translations decorator into a utility module
